requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
altgraph==0.17.4
anyio==4.3.0
APScheduler==3.9.1.post1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.15.0
backports.tarfile==1.2.0
bidict==0.23.1
boto3==1.34.105
botocore==1.34.105
build==1.2.1
CacheControl==0.12.14
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cohere==5.9.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confluent-kafka==2.4.0
cryptography==42.0.7
cycler==0.12.1
dask==2022.10.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.1
fastavro==1.9.7
fbmessenger==6.0.0
filelock==3.16.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jax==0.4.28
jeepney==0.8.0
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.17.3
keras==2.12.0
keyring==25.2.1
kiwisolver==1.4.5
langchain-core==0.2.34
langchain-ollama==0.1.1
langsmith==0.1.101
libclang==18.1.1
locket==1.0.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.5.3
mattermostwrapper==2.2
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.3.0
msgpack==1.0.8
multidict==5.2.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==2.6.3
nh3==0.2.18
numpy==1.23.5
oauthlib==3.2.2
ollama==0.3.1
openai==1.35.12
opt-einsum==3.3.0
orjson==3.10.6
packaging==21.3
pamqp==3.2.1
panda==0.3.1
pandas==2.2.2
parameterized==0.9.0
partd==1.4.2
pillow==10.3.0
pkginfo==1.10.0
pluggy==1.5.0
portalocker==2.8.2
prompt-toolkit==3.0.28
protobuf==4.23.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.23.3
pydot==1.4.2
Pygments==2.18.0
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.11
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.2
PyPDF2==3.0.1
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2022.7.1
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
readme_renderer==43.0
redis==4.6.0
regex==2022.10.31
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rocketchat-API==1.30.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
SecretStorage==3.3.3
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.27.1
sniffio==1.3.1
SQLAlchemy==1.4.52
structlog==23.3.0
structlog-sentry==2.1.0
tabulate==0.9.0
tarsafe==0.0.4
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tokenizers==0.20.0
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
twilio==8.2.2
twine==5.1.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-utils==0.1.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.18
uvloop==0.19.0
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==3.0.3
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
